Добавляет наставника
