+++ js/data.js
@@ -1,0 +1,44 @@
+const USERS = [
+  'user01',
+  'user02',
+  'user03',
+  'user04',
+  'user05',
+  'user06',
+  'user07',
+  'user08',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+
+const OBJECT_COUNT = 10;
+
+export {USERS, TYPES, TIME, PHOTOS, FEATURES, OBJECT_COUNT};
